feat: improve bucket endpoints

+++ apps/api/src/bucket/services/src/bucket.service.ts
@@ -1,4 +1,4 @@
-import {Inject, Injectable, Optional} from "@nestjs/common";
+import {Inject, Injectable, NotFoundException, Optional} from "@nestjs/common";
 import {Default, Validator} from "@spica-server/core/schema";
 import {
   BaseCollection,
@@ -133,6 +133,9 @@ export class BucketService extends BaseCollection<Bucket>("buckets") {
 
   async drop(id: string | ObjectId) {
     const schema = await super.findOneAndDelete({_id: new ObjectId(id)});
+    if (!schema) {
+      throw new NotFoundException(`Bucket with ID ${id} does not exist.`);
+    }
     await this.db.dropCollection(getBucketDataCollection(id));
     return schema;
   }
